refactor: improve code structure and readability
Refactored the codebase to enhance maintainability and readability without changing the existing functionality.
